Word List 2 (not ready for use)
Added a version of the word list that will more closely match the arrangement of types on actual cards-it's as a separate file because it still needs to be implemented, the scripts all still use the old version.
Also added some scripts that will let it store types you've put in like magic.mse-game does, though right now they do it crudely.